requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
numpy>=1.26.0
pandas>=2.2.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
python-slugify==8.0.1
tqdm==4.66.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-playwright==0.4.3 